+++ apps/storybook/.storybook/preview.tsx
@@ -1,15 +1,14 @@
 import React from "react";
-import type { Preview } from "@storybook/react";
 import { withThemeByDataAttribute } from "@storybook/addon-themes";
-import { customViewports } from "./screenSizes";
-import {OrderlyProvider} from "./orderlyProvider";
+import type { Preview } from "@storybook/react-vite";
 import { withThemeBuilder } from "../src/addons/theme_tool/preview";
-
+import { OrderlyProvider } from "../src/components/orderlyProvider";
+import { customViewports } from "./screenSizes";
 import "../src/tailwind.css";
 
 const preview: Preview = {
   decorators: [
-    (Story: any, context: any) => {
+    (Story) => {
       return (
         <OrderlyProvider>
           <Story />
@@ -17,6 +16,7 @@ const preview: Preview = {
       );
     },
     withThemeBuilder,
+    // https://github.com/storybookjs/storybook/blob/next/code/addons/themes/docs/getting-started/tailwind.md#%EF%B8%8F-using-a-data-attribute-for-theme
     withThemeByDataAttribute({
       themes: {
         orderly: "orderly",
@@ -29,18 +29,30 @@ const preview: Preview = {
   ],
   tags: ["autodocs"],
   parameters: {
-    backgrounds: {
-      values: [
-        { name: "Dark", value: "rgb(var(--oui-color-base-10))" },
-        { name: "Light", value: "rgb(255,255,255)" },
-      ],
-      default: "Dark",
-    },
+    // custom background colors: http://localhost:6008/?path=/story/package-trading-leaderboard--campaign-leaderboard&globals=theme:custom
+    // backgrounds: {
+    //   options: [
+    //     { name: "Dark", value: "rgb(var(--oui-color-base-10))" },
+    //     { name: "Light", value: "rgb(255,255,255)" },
+    //   ],
+    //   default: "Dark",
+    // },
     controls: {
       matchers: {
         color: /(background|color)$/i,
         date: /Date$/i,
       },
+    },
+
+    a11y: {
+      // 'todo' - show a11y violations in the test UI only
+      // 'error' - fail CI on a11y violations
+      // 'off' - skip a11y checks entirely
+      test: "todo",
+    },
+
+    docs: {
+      codePanel: true,
     },
     options: {
       storySort: {
@@ -71,12 +83,7 @@ const preview: Preview = {
       },
     },
     viewport: {
-      viewports: {
-        // ...INITIAL_VIEWPORTS,
-        // ...MINIMAL_VIEWPORTS,
-        ...customViewports,
-      },
-      // defaultViewport: 'iphone6 PRO',
+      options: customViewports,
     },
   },
 };