+++ apps/storybook/src/addons/theme_tool/components/radius.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { styled } from "@storybook/theming";
-import { useTheme } from "./editorContext";
+import { styled } from "storybook/theming";
+import { useTheme } from "./context";
 
 export const Wrap = styled.div`
   display: flex;