+++ apps/storybook/src/stories/customize/components2.stories.tsx
@@ -1,4 +1,5 @@
-import type { StoryObj } from "@storybook/react";
+import type { StoryObj } from "@storybook/react-vite";
+import { fn } from "storybook/test";
 import { OrderlyAppProvider } from "@orderly.network/react-app";
 import {
   Flex,
@@ -7,8 +8,6 @@ import {
   installExtension,
 } from "@orderly.network/ui";
 import { Scaffold } from "@orderly.network/ui-scaffold";
-import { fn } from "@storybook/test";
-
 import { WalletConnectorProvider } from "@orderly.network/wallet-connector";
 
 const meta = {