+++ packages/ui-transfer/src/components/convertForm/convertForm.script.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { useEffect, useMemo, useState } from "react";
 import {
-  useChains,
   useConfig,
   useConvert,
   useComputedLTV,
@@ -9,6 +8,7 @@ import {
   useOdosQuote,
   useWalletConnector,
   useWalletSubscription,
+  useMemoizedFn,
 } from "@orderly.network/hooks";
 import { useTranslation } from "@orderly.network/i18n";
 import { account } from "@orderly.network/perp";
@@ -16,9 +16,7 @@ import { useAppContext } from "@orderly.network/react-app";
 import { nativeETHAddress, nativeTokenAddress } from "@orderly.network/types";
 import type { API, NetworkId } from "@orderly.network/types";
 import { toast } from "@orderly.network/ui";
-import { Decimal, praseChainIdToNumber } from "@orderly.network/utils";
-import { InputStatus } from "../../types";
-import type { CurrentChain } from "../depositForm/hooks";
+import { Decimal } from "@orderly.network/utils";
 import { useSettlePnl } from "../unsettlePnlInfo/useSettlePnl";
 import { useToken } from "./hooks/useToken";
 
@@ -50,58 +48,29 @@ export const useConvertFormScript = (options: ConvertFormScriptOptions) => {
 
   const config = useConfig();
 
-  const brokerName = config.get("brokerName");
   const networkId = config.get("networkId") as NetworkId;
 
   const [quantity, setQuantity] = useState<string>("");
-  const [inputStatus, setInputStatus] = useState<InputStatus>("default");
-  const [hintMessage, setHintMessage] = useState<string>();
 
   const { wrongNetwork } = useAppContext();
 
-  const [, { findByChainId }] = useChains(networkId, {
-    pick: "network_infos",
-    filter: (chain: any) =>
-      chain.network_infos?.bridge_enable || chain.network_infos?.bridgeless,
-  });
+  const { wallet } = useWalletConnector();
 
-  const [linkDeviceStorage] = useLocalStorage("orderly_link_device", {});
+  const {
+    sourceToken,
+    sourceTokens,
+    onSourceTokenChange,
+    targetToken,
+    chainId,
+  } = useToken({ defaultValue: defaultToken });
 
-  const { connectedChain, wallet } = useWalletConnector();
-
-  const currentChain = useMemo(() => {
-    // if (!connectedChain) return null;
-
-    const chainId = connectedChain
-      ? praseChainIdToNumber(connectedChain.id)
-      : Number.parseInt(linkDeviceStorage?.chainId);
-
-    if (!chainId) {
-      return null;
-    }
-
-    const chain = findByChainId(chainId);
-
-    return {
-      ...connectedChain,
-      id: chainId,
-      info: chain!,
-    } as CurrentChain;
-  }, [findByChainId, connectedChain, linkDeviceStorage]);
-
-  const { sourceToken, sourceTokens, onSourceTokenChange, targetToken } =
-    useToken(
-      { currentChain, defaultValue: defaultToken },
-      (token) => token.symbol === "USDC" || token.is_collateral,
-    );
-
-  const token = useMemo<API.TokenInfo>(() => {
+  const token = useMemo<API.Chain>(() => {
     const _token = {
       ...sourceToken!,
-      precision: sourceToken?.precision ?? 6,
+      precision: sourceToken?.decimals ?? 6,
     };
     if (
-      _token.symbol === "ETH" &&
+      _token.token === "ETH" &&
       (!_token.address || _token.address === nativeTokenAddress)
     ) {
       _token.address = nativeETHAddress;
@@ -126,35 +95,10 @@ export const useConvertFormScript = (options: ConvertFormScriptOptions) => {
     1,
   );
 
-  const checkIsBridgeless = useMemo(() => {
-    if (wrongNetwork) {
-      return false;
-    }
-    if (!currentChain) {
-      return false;
-    }
-    if (networkId === "testnet") {
-      return true;
-    }
-    if (!currentChain.info) {
-      return false;
-    }
-    if (
-      !currentChain.info.network_infos ||
-      !currentChain.info.network_infos.bridgeless
-    ) {
-      return false;
-    }
-    return true;
-  }, [currentChain, wrongNetwork]);
-
-  const { maxAmount, convert } = useConvert({ token: token.symbol });
+  const { maxAmount, convert } = useConvert({ token: token.token });
 
   const onConvert = async () => {
     if (loading) {
-      return;
-    }
-    if (inputStatus !== "default") {
       return;
     }
     setLoading(true);
@@ -182,6 +126,8 @@ export const useConvertFormScript = (options: ConvertFormScriptOptions) => {
   const [postQuote, { data: quoteData, isMutating: isQuoteLoading }] =
     useOdosQuote();
 
+  const memoizedPostQuote = useMemoizedFn(postQuote);
+
   const convertRate = useMemo(() => {
     if (!quoteData || isQuoteLoading) {
       return "-";
@@ -192,30 +138,27 @@ export const useConvertFormScript = (options: ConvertFormScriptOptions) => {
       .div(unnormalizeAmount(quoteData.inAmounts[0], token.decimals))
       .toString();
     return rate;
-  }, [quoteData]);
+  }, [isQuoteLoading, quoteData, targetToken?.decimals, token?.decimals]);
 
   useEffect(() => {
-    if (quantity && currentChain?.id && token.address && targetToken?.address) {
-      postQuote({
-        // chainId: 8453,
-        chainId: currentChain.id,
+    if (quantity && chainId && token.address && targetToken?.address) {
+      memoizedPostQuote({
+        chainId: chainId,
         inputTokens: [
           {
             amount: normalizeAmount(quantity, token.decimals),
-            // tokenAddress: "0xfde4C96c8593536E31F229EA8f37b2ADa2699bb2",
             tokenAddress: token.address,
           },
         ],
         outputTokens: [
           {
             proportion: 1,
-            // tokenAddress: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
             tokenAddress: targetToken.address,
           },
         ],
       });
     }
-  }, [quantity, currentChain?.id, token, targetToken]);
+  }, [quantity, token, targetToken, chainId, memoizedPostQuote]);
 
   const minimumReceived = useMemo(() => {
     if (!quoteData || isQuoteLoading) {
@@ -231,20 +174,16 @@ export const useConvertFormScript = (options: ConvertFormScriptOptions) => {
 
   const nextLTV = useComputedLTV({
     input: Number(quantity),
-    token: sourceToken?.symbol,
+    token: sourceToken?.token,
   });
 
-  const disabled =
-    !quantity ||
-    Number(quantity) === 0 ||
-    ["error", "warning"].includes(inputStatus);
+  const disabled = !quantity || Number(quantity) === 0;
 
   useWalletSubscription({
     onMessage(data: any) {
       if (!crossChainTrans) {
         return;
       }
-
       const { trxId, transStatus } = data;
       if (trxId === crossChainTrans && transStatus === "COMPLETED") {
         setCrossChainTrans(false);
@@ -263,8 +202,6 @@ export const useConvertFormScript = (options: ConvertFormScriptOptions) => {
     sourceTokens,
     onSourceTokenChange,
     targetToken,
-    inputStatus,
-    hintMessage,
     balanceRevalidating: false,
     maxQuantity: maxAmount,
     disabled,
@@ -272,11 +209,9 @@ export const useConvertFormScript = (options: ConvertFormScriptOptions) => {
     wrongNetwork,
     onConvert,
     crossChainTrans,
-    networkId,
-    checkIsBridgeless,
     hasPositions,
     onSettlePnl,
-    brokerName,
+    networkId,
     slippage,
     onSlippageChange: setSlippage,
     convertRate,