+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,77 @@
 # storybook2
 
+## 1.5.1
+
+### Patch Changes
+
+- publish
+- 5ad8a1b: publish
+- Updated dependencies
+- Updated dependencies [5ad8a1b]
+  - @orderly.network/affiliate@2.5.1
+  - @orderly.network/react-app@2.5.1
+  - @orderly.network/chart@2.5.1
+  - @orderly.network/core@2.5.1
+  - @orderly.network/default-evm-adapter@2.5.1
+  - @orderly.network/default-solana-adapter@2.5.1
+  - @orderly.network/hooks@2.5.1
+  - @orderly.network/i18n@2.5.1
+  - @orderly.network/markets@2.5.1
+  - @orderly.network/portfolio@2.5.1
+  - @orderly.network/trading@2.5.1
+  - @orderly.network/trading-leaderboard@2.5.1
+  - @orderly.network/trading-rewards@2.5.1
+  - @orderly.network/types@2.5.1
+  - @orderly.network/ui@2.5.1
+  - @orderly.network/ui-chain-selector@2.5.1
+  - @orderly.network/ui-connector@2.5.1
+  - @orderly.network/ui-leverage@2.5.1
+  - @orderly.network/ui-order-entry@2.5.1
+  - @orderly.network/ui-orders@2.5.1
+  - @orderly.network/ui-positions@2.5.1
+  - @orderly.network/ui-scaffold@2.5.1
+  - @orderly.network/ui-share@2.5.1
+  - @orderly.network/ui-tpsl@2.5.1
+  - @orderly.network/ui-tradingview@2.5.1
+  - @orderly.network/utils@2.5.1
+  - @orderly.network/wallet-connector@2.5.1
+  - @orderly.network/wallet-connector-privy@2.5.1
+
+## 1.5.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/affiliate@2.5.1-alpha.0
+  - @orderly.network/react-app@2.5.1-alpha.0
+  - @orderly.network/chart@2.5.1-alpha.0
+  - @orderly.network/core@2.5.1-alpha.0
+  - @orderly.network/default-evm-adapter@2.5.1-alpha.0
+  - @orderly.network/default-solana-adapter@2.5.1-alpha.0
+  - @orderly.network/hooks@2.5.1-alpha.0
+  - @orderly.network/i18n@2.5.1-alpha.0
+  - @orderly.network/markets@2.5.1-alpha.0
+  - @orderly.network/portfolio@2.5.1-alpha.0
+  - @orderly.network/trading@2.5.1-alpha.0
+  - @orderly.network/trading-leaderboard@2.5.1-alpha.0
+  - @orderly.network/trading-rewards@2.5.1-alpha.0
+  - @orderly.network/types@2.5.1-alpha.0
+  - @orderly.network/ui@2.5.1-alpha.0
+  - @orderly.network/ui-chain-selector@2.5.1-alpha.0
+  - @orderly.network/ui-connector@2.5.1-alpha.0
+  - @orderly.network/ui-leverage@2.5.1-alpha.0
+  - @orderly.network/ui-order-entry@2.5.1-alpha.0
+  - @orderly.network/ui-orders@2.5.1-alpha.0
+  - @orderly.network/ui-positions@2.5.1-alpha.0
+  - @orderly.network/ui-scaffold@2.5.1-alpha.0
+  - @orderly.network/ui-share@2.5.1-alpha.0
+  - @orderly.network/ui-tpsl@2.5.1-alpha.0
+  - @orderly.network/ui-tradingview@2.5.1-alpha.0
+  - @orderly.network/utils@2.5.1-alpha.0
+  - @orderly.network/wallet-connector@2.5.1-alpha.0
+  - @orderly.network/wallet-connector-privy@2.5.1-alpha.0
+
 ## 1.5.0
 
 ### Minor Changes